danya-tattoo-voronezh/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
requests==2.32.3
pydantic-settings==2.7.0
